requirements.txt
djangorestframework
markdown
django-filter
mysqlclient
pillow
coreapi
django-guardian
